Updated BUILD file
Split tests into test_compilation and test_learning.py to work with BAZEL build files.
